# perf(ext/node): optimize fs.exists[Sync]