--- a/ext/fs/std_fs.rs
+++ b/ext/fs/std_fs.rs
@@ -784,16 +784,16 @@ fn exists(path: &Path) -> bool {
 
   #[cfg(windows)]
   {
+    use std::os::windows::ffi::OsStrExt;
     use winapi::um::fileapi::GetFileAttributesW;
-    use winapi::um::winnt::FILE_ATTRIBUTE_DIRECTORY;
-    use winapi::um::winnt::FILE_ATTRIBUTE_NORMAL;
-    use winapi::um::winnt::INVALID_FILE_ATTRIBUTES;
+    use winapi::um::fileapi::INVALID_FILE_ATTRIBUTES;
 
     let path = path
       .as_os_str()
       .encode_wide()
       .chain(std::iter::once(0))
       .collect::<Vec<_>>();
+    // Safety: `path` is a null-terminated string
     let attrs = unsafe { GetFileAttributesW(path.as_ptr()) };
 
     attrs != INVALID_FILE_ATTRIBUTES